+++ src/LaravelApiUserProviderServiceProvider.php
@@ -2,6 +2,11 @@
 
 namespace Amin3536\LaravelApiUserProvider;
 
+use App\Modules\authService\CustomTokenGuard;
+use App\Modules\authService\Deserializer;
+use App\Modules\authService\ExternalUserProvider;
+use App\Modules\interactModule\GuzzelHttpClient;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\ServiceProvider;
 
 class LaravelApiUserProviderServiceProvider extends ServiceProvider
@@ -11,17 +16,54 @@ class LaravelApiUserProviderServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function boot(): void
-    {
-        // $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'amin3536');
-        // $this->loadViewsFrom(__DIR__.'/../resources/views', 'amin3536');
-        // $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
-        // $this->loadRoutesFrom(__DIR__.'/routes.php');
+//    public function boot(): void
+//    {
+//        // $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'amin3536');
+//        // $this->loadViewsFrom(__DIR__.'/../resources/views', 'amin3536');
+//        // $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+//        // $this->loadRoutesFrom(__DIR__.'/routes.php');
+//
+//        // Publishing is only necessary when using the CLI.
+//
+//    }
 
-        // Publishing is only necessary when using the CLI.
-        if ($this->app->runningInConsole()) {
-            $this->bootForConsole();
-        }
+    public function boot()
+    {
+//       TODO: manage all service client  (auth , leitner   with container later   :)
+//        $this->app->bind(HttpClient::class, function ($app) {
+//            return $app->make(new GuzzelHttpClient());
+//        });
+
+        Auth::provider('api-provider', function ($app, array $config) {
+            return new ExternalUserProvider(new GuzzelHttpClient($this->getBaseUrl()), $config['model'], $config['url'],new Deserializer());
+        });
+
+        Auth::Extend('api-token', function ($app, $name, array $config) {
+            $guard = new CustomTokenGuard(
+                Auth::createUserProvider($config['provider']),
+                $app['request'],
+                $config['input_key'] ?? 'api_token',
+                $config['storage_key'] ?? 'api_token',
+                $config['hash'] ?? false
+            );
+            $this->app->refresh('request', $guard, 'setRequest');
+            return $guard;
+        });
+
+//        if ($this->app->runningInConsole()) {
+//            $this->bootForConsole();
+//        }
+    }
+
+
+    /**
+     * Get the user provider configuration.
+     *
+     * @return string|null
+     */
+    protected function getBaseUrl()
+    {
+        return $this->app['config']['auth.base-url'];
     }
 
     /**
@@ -31,12 +73,12 @@ class LaravelApiUserProviderServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        $this->mergeConfigFrom(__DIR__.'/../config/laravel-api-user-provider.php', 'laravel-api-user-provider');
+//        $this->mergeConfigFrom(__DIR__.'/../config/laravel-api-user-provider.php', 'laravel-api-user-provider');
 
-        // Register the service the package provides.
-        $this->app->singleton('laravel-api-user-provider', function ($app) {
-            return new LaravelApiUserProvider;
-        });
+//        // Register the service the package provides.
+//        $this->app->singleton('laravel-api-user-provider', function ($app) {
+//            return new LaravelApiUserProvider;
+//        });
     }
 
     /**
