+++ src/authService/Deserializer.php
@@ -9,6 +9,7 @@
 namespace App\Modules\authService;
 
 
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 
 class Deserializer implements DeserializerInterface
@@ -18,10 +19,6 @@ class Deserializer implements DeserializerInterface
      */
     private $model;
 
-    public function __construct(Model $model)
-    {
-        $this->model = $model;
-    }
 
 
     /**
@@ -50,8 +47,8 @@ class Deserializer implements DeserializerInterface
     }
 
     /**
-     * @param $roles
-     * @return \Illuminate\Database\Eloquent\Collection
+     * @param $list
+     * @return \Illuminate\Database\Eloquent\Collection|null
      */
     protected function convertList($list): Collection
     {
