+++ src/authService/DeserializerInterface.php
@@ -14,11 +14,7 @@ use Illuminate\Database\Eloquent\Model;
 interface DeserializerInterface
 {
 
-    /**
-     * DeserializerInterface constructor.
-     * @param Model $model
-     */
-    public function __construct(Model $model);
+
 
 
     /**
